Inline command server websocket code
Having this in a separate necessitates some confusing circular exports, it is much more straightforward to simply have this in the same file.
