+++ packages/server/src/command-server.ts
@@ -1,6 +1,6 @@
-import { Operation, spawn } from 'effection';
+import { Operation, spawn, fork } from 'effection';
 import { Mailbox } from '@bigtest/effection';
-import { express } from '@bigtest/effection-express';
+import { express, Socket } from '@bigtest/effection-express';
 import * as graphqlHTTP from 'express-graphql';
 import { graphql as executeGraphql } from 'graphql';
 
@@ -8,7 +8,7 @@ import { schema } from './schema';
 import { Atom } from '@bigtest/atom';
 import { OrchestratorState } from './orchestrator/state';
 
-import { handleMessage } from './command-server/websocket';
+import { Message, QueryMessage, MutationMessage, isQuery, isMutation } from './protocol';
 
 export type CommandMessage = { status: "ready" } | { type: "run"; id: string };
 
@@ -62,4 +62,45 @@ export function graphqlOptions(delegate: Mailbox, state: OrchestratorState) {
     rootValue: state,
     context: { delegate, testRunIds }
   };
+}
+
+export function handleMessage(delegate: Mailbox, atom: Atom<OrchestratorState>): (socket: Socket) => Operation {
+  function* handleQuery(message: QueryMessage, socket: Socket): Operation {
+    yield publishQueryResult(message, atom.get(), socket);
+
+    if (message.live) {
+      yield fork(subscribe(message, socket));
+    }
+  }
+
+  function* handleMutation(message: MutationMessage, socket: Socket): Operation {
+    let result = yield graphql(message.mutation, delegate, atom.get());
+    result.responseId = message.responseId;
+    yield socket.send(result);
+  }
+
+  function* publishQueryResult(message: QueryMessage, state: OrchestratorState, socket: Socket): Operation {
+    let result = yield graphql(message.query, delegate, state);
+    result.responseId = message.responseId;
+    yield socket.send(result);
+  }
+
+  function* subscribe(message: QueryMessage, socket: Socket) {
+    yield atom.each(state => publishQueryResult(message, state, socket));
+  }
+
+  return function*(socket) {
+    let messages: Mailbox = yield socket.subscribe();
+
+    while (true) {
+      let message: Message = yield messages.receive();
+
+      if (isQuery(message)) {
+        yield fork(handleQuery(message, socket));
+      }
+      if (isMutation(message)) {
+        yield fork(handleMutation(message, socket));
+      }
+    }
+  }
 }
